LiveAI/requirements.txt
google-generativeai
duckduckgo-search
faster-whisper
ctranslate2
edge-tts
faiss-cpu
sentence-transformers
sounddevice
scipy
numpy